+++ src/heretic/r_data.cpp
@@ -336,14 +336,14 @@ void R_InitTextures()
 //
     maptex = maptex1 = cache_lump_name<int *>(texture1, PU_STATIC);
     numtextures1 = LONG(*maptex);
-    maxoff = W_LumpLength(W_GetNumForName(texture1));
+    maxoff = static_cast<int>(W_LumpLength(W_GetNumForName(texture1)));
     directory = maptex + 1;
 
     if (W_CheckNumForName(texture2) != -1)
     {
         maptex2 = cache_lump_name<int *>(texture2, PU_STATIC);
         numtextures2 = LONG(*maptex2);
-        maxoff2 = W_LumpLength(W_GetNumForName(texture2));
+        maxoff2 = static_cast<int>(W_LumpLength(W_GetNumForName(texture2)));
     }
     else
     {
@@ -528,7 +528,7 @@ void R_InitColormaps()
 // 256 byte align tables
 //
     lump = W_GetNumForName(DEH_String("COLORMAP"));
-    length = W_LumpLength(lump);
+    length = static_cast<int>(W_LumpLength(lump));
     colormaps = zmalloc<lighttable_t *>(length, PU_STATIC, 0);
     W_ReadLump(lump, colormaps);
 }
@@ -671,7 +671,7 @@ void R_PrecacheLevel()
         if (flatpresent[i])
         {
             lump = firstflat + i;
-            flatmemory += lumpinfo[lump]->size;
+            flatmemory += static_cast<int>(lumpinfo[lump]->size);
             W_CacheLumpNum(lump, PU_CACHE);
         }
 
@@ -701,7 +701,7 @@ void R_PrecacheLevel()
         for (j = 0; j < texture->patchcount; j++)
         {
             lump = texture->patches[j].patch;
-            texturememory += lumpinfo[lump]->size;
+            texturememory += static_cast<int>(lumpinfo[lump]->size);
             W_CacheLumpNum(lump, PU_CACHE);
         }
     }
@@ -732,7 +732,7 @@ void R_PrecacheLevel()
             for (k = 0; k < 8; k++)
             {
                 lump = firstspritelump + sf->lump[k];
-                spritememory += lumpinfo[lump]->size;
+                spritememory += static_cast<int>(lumpinfo[lump]->size);
                 W_CacheLumpNum(lump, PU_CACHE);
             }
         }
