+++ src/heretic/s_sound.cpp
@@ -95,7 +95,7 @@ void S_StartSong(int song, bool loop)
     }
     mus_lumpnum = W_GetNumForName(S_music[song].name);
     mus_sndptr = cache_lump_num<void *>(mus_lumpnum, PU_MUSIC);
-    mus_len = W_LumpLength(mus_lumpnum);
+    mus_len = static_cast<int>(W_LumpLength(mus_lumpnum));
     rs = I_RegisterSong(mus_sndptr, mus_len);
     I_PlaySong(rs, loop);       //'true' denotes endless looping.
     mus_song = song;
