+++ src/i_timer.cpp
@@ -55,14 +55,14 @@ int I_GetTimeMS()
     if (basetime == 0)
         basetime = ticks;
 
-    return ticks - basetime;
+    return static_cast<int>(ticks - basetime);
 }
 
 // Sleep for a specified number of ms
 
 void I_Sleep(int ms)
 {
-    SDL_Delay(ms);
+    SDL_Delay(static_cast<Uint32>(ms));
 }
 
 void I_WaitVBL(int count)
